tracer: treat IDs as 64-bit signed ints
Jaeger defines that trace and span IDs are `i64`:
https://github.com/jaegertracing/jaeger-idl/blob/master/thrift/jaeger.thrift

In Thrift, `i64` are 64-bit signed integers:
https://thrift.apache.org/docs/types

The code was simply calling `to_i(16)` on the IDs, the result of
that is a 64-bit unsigned integer. This causes problems in Thrift when
sending values outside the range that can be represented using 64-bit
signed integers.

This commit solves the issue by making sure that both the span ID and the
parent span ID are interpreted as 64-bit signed integers.