Fix transforming big unsigned ints to thrift signed ints
Conversion methods are now explicit and hopefully more understandable.
Previously values over 2**63 were not handled correctly. The confusion
came because id values from http headers are unsigned int64 but the
thrift span is expecting signed int64. Made sure that internally we're
always using uint64 and we'll convert this to int64 only when creating a
thrift span.

Closes #12