Return sample result and tags when calling Sampler#sample?
Previously we returned only a boolean value which indicated if the trace
should be sampled or not. We had a separate method for getting the tags.
This doesn't allow creating samplers that can return different tags.
It's needed to implement GuaranteedThroughputProbabilistic sampler.

The sampling check itself was also moved from create_parent_context
method to Tracer. This allows providing more information to the
samplers. E.g. now we also pass operation_name to sampler.